case2/requirements.txt
pytest==5.2.2
pytest-randomly==3.1.0
pytest-xdist==1.30.0
PyYAML==5.1.2
selenium==3.141.0
